+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 # IBAN-Generator Changelog
 
 ## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+* Generate a Random French Iban in a ToolWindow